Refactoring: base & compilation classes
