+++ app/src/main/java/saarland/cispa/artist/artistgui/compilation/CompilationPresenter.java
@@ -33,8 +33,6 @@ import java.io.FileOutputStream;
 import java.io.IOException;
 
 import saarland.cispa.artist.ArtistImpl;
-import saarland.cispa.artist.artistgui.CompilationService;
-import saarland.cispa.artist.artistgui.CompileDialogActivity;
 import saarland.cispa.artist.artistgui.MainActivity;
 import saarland.cispa.artist.artistgui.compilation.notification.CompileNotificationManager;
 import saarland.cispa.artist.artistgui.settings.config.ArtistAppConfig;
